* adjust bitset and compactarray for compatibility
